feat: Add new migrations, modify checkout models
1.New migration files were added for the `profiles` and `checkout` apps
to update their models. This includes creating the UserProfile model &
adding new fields to the Order model for stripe PID and original cart.

2. Modifications were made to the `checkout` app to accommodate changes
in the order management system. The Order model now uses CountryField
instead of CharField for the 'country' field, and it now includes
'stripe_pid' and 'original_cart'. The `views.py` file in the `checkout`
app saves 'stripe_pid' and 'original_cart' when an order is created.

These changes improve the checkout process and enhance order tracking by
including the stripe payment ID and the state of the cart at the time of
the order. The checkout admin view now displays 'original_cart' and
'stripe_pid', and the checkout form no longer includes 'country' in its
placeholders.

3.Aesthetic improvements were made to `checkout.html`,
including the option to save delivery information for authenticated
users.

summary: these updates make the checkout process more reliable and
improve the user experience, particularly for registered users.