feat: add Stripe webhook for successful payment
1. Handles an additional Stripe Webhook for the event payment_intent
.succeeded. The changes enrich the payment flow, checking if an order
exists in the database, if the order does not exist, it creates a new
order with line items based on the cart information from the payment
intent. Improves data validation for empty fields in shipping
details. This is crucial to ensure the app's robustness, preventing
potential errors and enhancing user experience.