Refactor: settings.py for production
This commit carries out two tasks. 

To adhere to PEP8 style guidelines, the import statements in settings.py
were rearranged. The import for `Path` was moved after `dj_database_url`,  and the unnecessary repeated import of `dj_database_url` in the 'else' block was removed.

In non-development mode, the value `ci-project-boutique-ado` was assigned to `ALLOWED_HOSTS`. This enhances security by restricting the application to specified host(s) in production mode.

To guide Heroku in starting the project, a new file named 'Procfile' was added. It instructs Heroku to run the gunicorn server with the core WSGI application.